requirements.txt
Django==3.2.9
django-crispy-forms==1.13.0
pycrypto==2.6.1